+++ familyhelper-ui-view/familyhelper-ui-view-pc/src/views/items/note/noteManagement/ItemEditPanel.vue
@@ -106,7 +106,7 @@ import { operateInspect, operatePut } from '@/api/settingrepo/settingNode';
 import { currentTimestamp, formatTimestamp } from '@/util/timestamp';
 import resolveResponse from '@/util/response';
 
-import { NOTE_ATTACHMENT_FILE } from '@/views/items/miscellaneous/fileEditor/filtTypeConstants';
+import { NOTE_ATTACHMENT_FILE } from '@/views/items/miscellaneous/fileEditor/fileTypeConstants';
 
 // noinspection JSAnnotator
 export default {